feature: cas.chat test

+++ ingus.php
@@ -12,7 +12,12 @@ global $bot;
 $bot = new \Ng\Ingus\Controller\Bot();
 $bot->get_updates( array() );
 global $cache;
-$cache    = new \Ng\Ingus\Controller\Cache( 'cache.json' );
+$cache = new \Ng\Ingus\Controller\Cache( 'cache.json' );
+if ( ! empty( $argv ) ) {
+	foreach ( $argv as $arg ) {
+		define( 'NGINS_ARGV_' . str_replace( '-', '', strtoupper( $arg ) ), true );
+	}
+}
 $rules    = json_decode( file_get_contents( __DIR__ . '/rules/rules.json' ), true );
 $cas_chat = new \Ng\Ingus\Controller\Cas_Chat( $cache );
 if ( ! $rules || ! is_array( $rules ) ) {
@@ -24,13 +29,23 @@ if ( empty( $bot->data['updates'] ) ) {
 	exit;
 }
 
-shuffle( $bot->data['updates'] );
+// shuffle( $bot->data['updates'] );
 foreach ( $bot->data['updates'] as $update ) {
+	// if user joined group
 	$update_id = $update->getUpdateId();
-	echo 'checking ' . $update_id . PHP_EOL;
 	if ( $cache->get( NG_ING_PREFIX . '_update_' . $update_id ) ) {
 		continue;
 	}
+	if ( $update->getMessage()->getNewChatMembers() && $cas_chat->check_user( $update->getMessage()->getFrom()->getId() ) ) {
+		send_message( NGINS_ADMIN_CHAT, 'User is banned in cas.chat, muted' . "\n" . $update_info );
+		$cache->set( NG_ING_PREFIX . '_update_' . $update_id, $update_id, 24 * 60 * 60 );
+		restrict_user( $update );
+		continue;
+	}
+	if ( $update->getMessage()->getLeftChatMember() ) {
+		continue;
+	}
+	echo 'checking ' . $update_id . PHP_EOL;
 	$update_info = $bot->get_update_info( $update );
 	$guard       = new \Ng\Ingus\Controller\Guard( $update );
 	if ( $cas_chat->check_user( $update->getMessage()->getFrom()->getId() ) ) {
@@ -66,7 +81,10 @@ function send_message( $chat_id, $text ) {
 	$bot->send_message( $chat_id, $text );
 }
 function restrict_user( $update ) {
-	return;
+	if ( defined( 'NGINS_ARGV_DRYRUN' ) && NGINS_ARGV_DRYRUN ) {
+		echo 'dry run' . PHP_EOL;
+		return;
+	}
 	global $bot;
 	$bot->restrict_chat_member( $update->getMessage()->getChat()->getId(), $update->getMessage()->getFrom()->getId() );
 	$bot->delete_message( $update->getMessage()->getChat()->getId(), (int) $update->getMessage()->getMessageId() );
